BT history: layout list
